+++ example/src/App.tsx
@@ -14,6 +14,7 @@ import {
   createTracker,
   disableTracking,
   enableTracking,
+  setCustomDimension,
   setLogger,
   setUserId,
   setVisitorId,
@@ -29,14 +30,14 @@ import {
   trackOutlink,
   trackScreen,
   trackSearch,
-} from '@logicwind/react-native-matomo-tracker';
+} from '@rausgegangen/react-native-matomo-tracker';
 
 export default function App() {
   const [result] = React.useState<number | undefined>();
 
   React.useEffect(() => {
-    createTracker("https://your-matomo-url/matomo.php", 1) //Replace 1 with your matomo site id
-    setLogger()
+    createTracker('https://your-matomo-url/matomo.php', 1); //Replace 1 with your matomo site id
+    setLogger();
   }, []);
 
   return (
@@ -48,12 +49,16 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-              trackCampaign("Home screen","https://rntestApp://home?mtm_campaign=2020_august_promo&mtm_source=google&mtm_medium=email&mtm_keyword=2020 august promo&mtm_content=primary-cta",[
-                {  key: "1",value:"visit dimension 1"},
-                  {  key:"2",value:"visit dimension 2"},
-                {    key:"3",value:"action dimension 3"},
-                 {  key:"4",value:"action dimension 4"},
-               ])
+              trackCampaign(
+                'Home screen',
+                'https://rntestApp://home?mtm_campaign=2020_august_promo&mtm_source=google&mtm_medium=email&mtm_keyword=2020 august promo&mtm_content=primary-cta',
+                [
+                  { key: '1', value: 'visit dimension 1' },
+                  { key: '2', value: 'visit dimension 2' },
+                  { key: '3', value: 'action dimension 3' },
+                  { key: '4', value: 'action dimension 4' },
+                ]
+              );
             }}
           >
             <Text style={styles.buttonText}>Track Campaign</Text>
@@ -62,7 +67,7 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-              startSession()
+              startSession();
             }}
           >
             <Text style={styles.buttonText}>Start Session</Text>
@@ -70,14 +75,12 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-              trackScreen('HomeScreen', 'This is test home screen',
-              [
-                 {  key: "1",value:"visit dimension 1"},
-                   {  key:"2",value:"visit dimension 2"},
-                 {    key:"3",value:"action dimension 3"},
-                  {  key:"4",value:"action dimension 4"},
-                ]
-              );
+              trackScreen('HomeScreen', 'This is test home screen', [
+                { key: '1', value: 'visit dimension 1' },
+                { key: '2', value: 'visit dimension 2' },
+                { key: '3', value: 'action dimension 3' },
+                { key: '4', value: 'action dimension 4' },
+              ]);
             }}
           >
             <Text style={styles.buttonText}>Track Screen</Text>
@@ -86,12 +89,12 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-              trackEvent('test category', 'test action', ' test name', 2,[
-                {  key: "1",value:"visit dimension 1"},
-                  {  key:"2",value:"visit dimension 2"},
-                {    key:"3",value:"action dimension 3"},
-                 {  key:"4",value:"action dimension 4"},
-               ]);
+              trackEvent('test category', 'test action', ' test name', 2, [
+                { key: '1', value: 'visit dimension 1' },
+                { key: '2', value: 'visit dimension 2' },
+                { key: '3', value: 'action dimension 3' },
+                { key: '4', value: 'action dimension 4' },
+              ]);
               trackEvent(
                 'basket',
                 JSON.stringify({
@@ -99,15 +102,15 @@ export default function App() {
                   item: 'mens grey t-shirt',
                   description: ['round neck', 'long sleeved'],
                   size: 'L',
-                },),
+                }),
                 '',
                 0,
                 [
-                  {  key: "1",value:"visit dimension 1"},
-                    {  key:"2",value:"visit dimension 2"},
-                  {    key:"3",value:"action dimension 3"},
-                   {  key:"4",value:"action dimension 4"},
-                 ]
+                  { key: '1', value: 'visit dimension 1' },
+                  { key: '2', value: 'visit dimension 2' },
+                  { key: '3', value: 'action dimension 3' },
+                  { key: '4', value: 'action dimension 4' },
+                ]
               );
             }}
           >
@@ -116,12 +119,12 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-              trackOutlink('https://www.google.com/',[
-                {  key: "1",value:"visit dimension 1"},
-                  {  key:"2",value:"visit dimension 2"},
-                {    key:"3",value:"action dimension 3"},
-                 {  key:"4",value:"action dimension 4"},
-               ]);
+              trackOutlink('https://www.google.com/', [
+                { key: '1', value: 'visit dimension 1' },
+                { key: '2', value: 'visit dimension 2' },
+                { key: '3', value: 'action dimension 3' },
+                { key: '4', value: 'action dimension 4' },
+              ]);
             }}
           >
             <Text style={styles.buttonText}>Track OutLink</Text>
@@ -130,12 +133,12 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-              trackSearch('Logicwind', [
-                {  key: "1",value:"visit dimension 1"},
-                  {  key:"2",value:"visit dimension 2"},
-                {    key:"3",value:"action dimension 3"},
-                 {  key:"4",value:"action dimension 4"},
-               ]);
+              trackSearch('Rausgegangen', [
+                { key: '1', value: 'visit dimension 1' },
+                { key: '2', value: 'visit dimension 2' },
+                { key: '3', value: 'action dimension 3' },
+                { key: '4', value: 'action dimension 4' },
+              ]);
             }}
           >
             <Text style={styles.buttonText}>Track Search</Text>
@@ -144,12 +147,12 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-              trackImpression('Test Track Impression', [
-                {  key: "1",value:"visit dimension 1"},
-                  {  key:"2",value:"visit dimension 2"},
-                {    key:"3",value:"action dimension 3"},
-                 {  key:"4",value:"action dimension 4"},
-               ]);
+              trackImpression('Test Track Impression', '', '', [
+                { key: '1', value: 'visit dimension 1' },
+                { key: '2', value: 'visit dimension 2' },
+                { key: '3', value: 'action dimension 3' },
+                { key: '4', value: 'action dimension 4' },
+              ]);
             }}
           >
             <Text style={styles.buttonText}>Track Impression</Text>
@@ -158,12 +161,18 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-              trackInteraction('Test Track interaction', 'test inetraction', [
-                {  key: "1",value:"visit dimension 1"},
-                  {  key:"2",value:"visit dimension 2"},
-                {    key:"3",value:"action dimension 3"},
-                 {  key:"4",value:"action dimension 4"},
-               ]);
+              trackInteraction(
+                'Test Track interaction',
+                'test inetraction',
+                '',
+                '',
+                [
+                  { key: '1', value: 'visit dimension 1' },
+                  { key: '2', value: 'visit dimension 2' },
+                  { key: '3', value: 'action dimension 3' },
+                  { key: '4', value: 'action dimension 4' },
+                ]
+              );
             }}
           >
             <Text style={styles.buttonText}>Track Interaction</Text>
@@ -177,11 +186,11 @@ export default function App() {
                 'PDF Download',
                 'https://example.com/download.pdf',
                 [
-                  {  key: "1",value:"visit dimension 1"},
-                    {  key:"2",value:"visit dimension 2"},
-                  {    key:"3",value:"action dimension 3"},
-                   {  key:"4",value:"action dimension 4"},
-                 ]
+                  { key: '1', value: 'visit dimension 1' },
+                  { key: '2', value: 'visit dimension 2' },
+                  { key: '3', value: 'action dimension 3' },
+                  { key: '4', value: 'action dimension 4' },
+                ]
               );
             }}
           >
@@ -209,7 +218,7 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-              disableTracking()
+              disableTracking();
             }}
           >
             <Text style={styles.buttonText}>Disable Tracking</Text>
@@ -218,7 +227,7 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-              enableTracking()
+              enableTracking();
             }}
           >
             <Text style={styles.buttonText}>Enable Tracking</Text>
@@ -227,7 +236,7 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-              setLogger()
+              setLogger();
             }}
           >
             <Text style={styles.buttonText}>Set Logger</Text>
@@ -236,14 +245,27 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-              trackMediaEvent({ siteId: "siteId", mediaId: Date.now.toString(), mediaTitle: "video media play track", playerName: "test 08", mediaType: MediaType.VIDEO, mediaResource: "http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4", mediaStatus: "100", mediaLength: "100", mediaFullScreen: "1", mediaHeight: "720", mediaWidth: "1080", mediaProgress: "100",
-              dimension:[
-                {  key: "1",value:"visit dimension 1"},
-                  {  key:"2",value:"visit dimension 2"},
-                {    key:"3",value:"action dimension 3"},
-                 {  key:"4",value:"action dimension 4"},
-               ]
-            });
+              trackMediaEvent({
+                siteId: 'siteId',
+                mediaId: Date.now.toString(),
+                mediaTitle: 'video media play track',
+                playerName: 'test 08',
+                mediaType: MediaType.VIDEO,
+                mediaResource:
+                  'http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4',
+                mediaStatus: '100',
+                mediaLength: '100',
+                mediaFullScreen: '1',
+                mediaHeight: '720',
+                mediaWidth: '1080',
+                mediaProgress: '100',
+                dimension: [
+                  { key: '1', value: 'visit dimension 1' },
+                  { key: '2', value: 'visit dimension 2' },
+                  { key: '3', value: 'action dimension 3' },
+                  { key: '4', value: 'action dimension 4' },
+                ],
+              });
             }}
           >
             <Text style={styles.buttonText}>Video Play Stop</Text>
@@ -252,15 +274,14 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-          
-              trackCustomDimension({ 
-              dimensions:[
-                {  key: "1",value:"visit dimension 1"},
-                  {  key:"2",value:"visit dimension 2"},
-                {    key:"3",value:"action dimension 3"},
-                 {  key:"4",value:"action dimension 4"},
-               ]
-            });
+              trackCustomDimension({
+                dimensions: [
+                  { key: '1', value: 'visit dimension 1' },
+                  { key: '2', value: 'visit dimension 2' },
+                  { key: '3', value: 'action dimension 3' },
+                  { key: '4', value: 'action dimension 4' },
+                ],
+              });
             }}
           >
             <Text style={styles.buttonText}>Track Custom Dimension</Text>
@@ -269,8 +290,16 @@ export default function App() {
           <Pressable
             style={styles.button}
             onPress={() => {
-              trackDispatch()
+              setCustomDimension(1, 'premium-user');
+            }}
+          >
+            <Text style={styles.buttonText}>Set Custom Dimension</Text>
+          </Pressable>
 
+          <Pressable
+            style={styles.button}
+            onPress={() => {
+              trackDispatch();
             }}
           >
             <Text style={styles.buttonText}>Track Dispatch</Text>